Set debug_view_attributes from layout inspector.
In order to get traces of the properties we need to set
debug_view_attributes via abd. Add code to do this from Studio.
We will set this for a single package or for all packages if the
app has multiple app modules.

The Layout inspector currently does not work for applications
with multiple app modules, and this CL does not attempt to fix
that.

Bug: 122274789
Test: N/A
Change-Id: Idef35272489d066645ef8fbbaab9a1f7b50c6b7b