Fix "Include Kotlin support" stickiness when project creation cancelled - IJ plugin
Follow up for Change-Id: I33156fad4f30da6a894946f26c78f877aa89ca36
Previous fix was not working when Android Studio is run as
an IJ plugin.

When AS runs as a plugin in IJ, a NewProjectModel() object is
created in com.android.tools.idea.npw.ideahost.AndroidModuleBuilder and
the listener to save the wizard state is not called.
This CL moves the responsibility of saving the state of the Model to
the Model itself, instead of depending on a external listener call.

Fixes: 75512603
Test: manual (build and run IDEA instead of Android Studio)
Change-Id: I4a40909b5cd489be49aa44551650f5e02acfad32