Moved sending build failure metrics to onEnd instead of onFailure
BuildOutputInstantReader waits on the parsers thread only on closing,
and since onFailure is called before the reader is closed this
introduced flakiness in the integration test. This change moves the
metrics sending to be done after the reader is closed, and adds another
test to the integration test so if the same problem happened again it
should fail every time instead of being flaky.

Fixes: 126882940
Test: BuildOutputParsersIntegrationTest
Change-Id: Iaf8a17a67c3342464544f55cf736b0e5db781bff