Do not remove the aspects agent log if it exists
We were deleting the log file (if it existed) before start logging the
tests just to make it more convenient when running the tests locally.
However, a side-effect of that is we delete the log every time we start
running the tests of a new TestGroup. As we run the UNRELIABLE TestGroup
after the DEFAULT one, we end up with the logs of the former in the log
file.

We are more interested in the violations of the DEFAULT test group, so
we need at least to preserve that. The cleanest way to do it is simply
not deleting the log before running the tests of a TestGroup, so we
preserve the violations of every group.

This means we should remove the log before running the tests locally,
which will be added to the instructions of how to verify the agent
results/output. When running in go/ab there is no issue since we create
a new logs dir every time, so there will be no stale log file.

Note: even though we include the violations of the UNRELIABLE TestGroup
in the logs, AspectsAgentLogTest will fail only if some tests in the
DEFAULT test group fail.

Bug: 128894267
Test: UI tests should pass normally.
Change-Id: I46db20193ce02536266f36562dc9441947832007