After completion, put cursor between brackets if method has parameters
The appending of brackets is now handled in InsertHandler. As a side
effect, I was forced to update the tests because lookup strings no
longer contain brackets.

I also reverted what my changes of cfd7b614. Turns out it made things
more messy, because now we have logic we want to test in InsertHandler.

Test: DataBindingCodeCompletionTest. 1 existing test modified to check
for caret. 1 new test.
Fixes: 129501142
Change-Id: I4ad9ef57da4ff6b08672ba28e10ff7d780d330ea