Warn when debug with 'undebuggable app and system image'
All emulator system images used to be debuggable, however,
when 'Play store image' was introduced, this's no longer true.
In general, in order to be able to debug, either app is
debuggable, or system image is debuggable. So as to detect
and warn users at early stage, 'has play store' from
avd info is fetched for launchable device. For connected
device, original logic of 'querying build type' is for
debuggable checking.

Bug: b/131170197
Test: n/a
Change-Id: I7ddcb0551ed2148da76f6e3614f644afdb773a98