[New Sync] Add project location to cached model.
The cached model is loaded from cache file under .idea/caches,
the current code only verifies the checksum and timestamp.
It is possible that a project is copied or moved to different
location, with the cache file copied as well.

The cache should not be used in this case because some models
contain location information, for example, the content entries.

This change addes project location to cached model to catch this
case. If the cached project location is different from current
project, then treat the cache as invalid and perform full sync.

Bug: 127848590
Test: testLoadCacheWithDifferentProjectRoot
Change-Id: Iea482f06cd1d5ccb6de9d0043fe7517fbba7c0e4