Fix RenameTest
It hardcoded an expected implementation of a RenameHandler that handles
renaming a Gradle source root, but this changed in a recent platform
change (f63f8a343fed3273fe40d3554acc6581d48daf59) which broke the test
if the first source root read from ModuleRootManager was a res/
directory.  Instead, we now checked a known source root and do what
RenameAction does, finding the first handler that can handle a given
PsiElement.

Test: RenameTest
Fixes: 129956954
Change-Id: I8f2b74b4f1c6023719f5dec929de26bc5dd364e9