Remove no longer necessary deferring of project setup
since AndroidGradleProjectOpenProcessor does not loner request sync
itself and at the point of setupProject the project is already
initialized now.

This is important since even when a project is initialized the startup
manager postpones setupProject with invokeLater. This in turn
introduces a race condition in IDEA sync completion which may result in
sync completing when the project is still empty.

Bug: 130736558
Bug: 131564247
Test: n/a (manually tested creating and opening/importing projects)
Change-Id: Ia5f028da4bdfb1a0995a4ed50562e27fd8990f34