Ensure aspects_agent_log.txt is created on time
Makes sure GuiTests.getGuiTestRootDirPath().getAbsolutePath()/system/log
is created before setting the ASPECTS_AGENT_LOG environment variable, so
the aspects agent can access the log.

Before this change, we were trying to set the ASPECTS_AGENT_LOG variable
before the logs directory was creating, consequently setting it to null.
That implies the first TestGroup would fail to be logged. In our case,
the first TestGroup to run is the DEFAULT one, which is exactly the one
that should be logged in order to fail the test.

Bug: 131603363
Test: UI tests should pass
Change-Id: Ic89d7ac4f6e12d6ed44b3581c5259ce845762b4d