Fix a ProjectImpl memory leak caused by callback to IntellijService
This is a two part CL. This first part is a surface level fix to help
free project references that can cause other parts of studio to
behave in an unexpected way. The second part will clean up the rest of
the profilers object also leaked by closing the project.

Test: Manual verification with yourkit.
Bug: 131356561
Change-Id: I4907983165221b6c1afecb830f0ffcc7da113a06