Fixes session display bug in unified pipeline
This change fixes a bug where old session data was shown in a new session
when the unified pipeline is active. This fix applies to the overview
stage only, and TODOs have been placed in areas that need to be
addressed for the detailed stages.

The cause of this bug was that under the new unified pipeline, if we use
the view range to query data, we'll get data from previous sessions as
long as they fit within the bounds of the view.

The fix involved adding a new intersectRange to RangedSeries. We use this
range to ensure that the old data that is within the viewRange(the default
range) but not within the dataRange(the intersecting range) does not get
returned. ActivityComponent and EventComponent have additional custom
clamping logic at the moment, but TODOs have been placed to address
this in follow-up changes.

bug: 122964201
Test: - onlyDrawsWithinViewRange in LineChartTest
      - Modified testNewPipelineGetData in MemoryUsageTest and
        DetailedMemoryUsageTest by setting the dataRange in the beginning so
        that it can correctly intersect with the view range.
Change-Id: I3d6e41fbcf0323d329c2de95930d7d606fc08954