Fix refresh issue in energy and network profilers
Before this CL, the TimelineTable would only refresh when
a row was added or removed. When you made a selection and
dragged it around, this could feel like subtle stuttering
in the network profiler, but the energy profiler's behavior
could be a lot worse, since often jobs are long lived and
you may not add or remove new rows for minutes at a time.

After this CL, the TimelineTable class now responds to the
selection range changing, at which point, it refreshes the
component.

Fixes: 131699348
Test: N/A
Change-Id: I7111e37208dcf17f4ba417b9ba98eead9f82a51c
(cherry picked from commit 58b73ed19e5b22404322f6b572eb93901bb03d7d)