Fixes race condition on no-change Run not owned by AndroidProcessHandler
When the remote debugger is attached or when the app is run from the device
directly, Running/Debugging an unchanged app may be fast enough that the delta
install "am force-stop" command's effect does not get reflected in ddmlib
before the same IDevice is added to AndroidProcessHandler.

This is caused by the fact that a no-change Run does not wait until the stale
Client is actually terminated before proceeding to attempt to connect the
AndroidProcessHandler to the desired device. In such circumstances, the
handler will connect to the stale Client, and almost immediately have the same
Client's killed state get reflected by ddmlib, and removed from the process
handler.

The fix is to ensure that all existing Clients matching the application ID are
fully terminated before proceeding with handoff to AndroidProcessHandler.

Fixes: 131722013
Test: Run no-change builds on remote debug sessions and ensure process killed
Change-Id: I5d78ba0972060ebac1a36ea94f62b195d65819a3