Close the combo popup before performing a bound action.
The only action used in the Nele properties panel is currently:
"More Fonts..." under fontFamily. Currently the popup from the
ComboBox is kept open while the font dialog is open.
That causes issues where the popup is diaplyed on top of other
applications even when the other application has focus.

Fix this by performing the action of opening the dialog "later"
i.e. posting an event to the AWT event queue and close the popup
immediately.

Fixes: 131630401
Test: Visual effect only.
Change-Id: I7e899f38a4c7e0bff72a335dbbd6eaffe97a4506