Make sure a new property can be specified with resource picker.
The resource picker action should use the delegate property instead
of itself when dealing with a new property in the declared section.

Do this by adding a "delegate" property on the base NelePropertyItem
that returns itself. A NeleNewPropertyItem will return the correct
delegated property to find active component(s) from.

Fixes: 131772104
Test: Visual only. In the value of a new item the reource item
      should be visible.
Change-Id: If2996c9fe1449fc305667be43d33acafe0b7a832