Close the margin popup when mouse is outside the popup/button.
Background: there is logic in IJ to support opening a popup
from a popup which will keep the first popup open until all
dependent popups have closed.

If we open the margin popup first there is no logic to close it
when the mouse moves out of the popup. An additional problem is
that the margin popup is opened below the button on the toolbar
and the user may never move the mouse over the popup.

This change will close the popup when the mouse is moved outside
the area consisting of the popup and the tool bar button or the
small space between them.

The toolbar button (ActionButton) is not available in the view
handler so we infer it from the source of a mouse event.

Fixes: 131775093
Test: Visual only.
Change-Id: Ie04b1e50d1676d4341e206f63a9932ac07a57dba