Replace the old settings with Version 3 once
In Commit 54bacc2cb9 I repair the broken settings only if they match the
old default (the developer never changed them). As the breakage is
significant (when a developer reformats their XML code the arrangement
engine rearranges elements as well which changes XML semantics in bad
ways) we decided to go further and completely clobber the old settings
with the fixed defaults once. This way we also cover developers that
changed the old defaults.

Bug: 129323345
Test: AndroidCodeStyleSettingsTest
  ::testInitializeSetsArrangementSettingsToVersion3Once
Change-Id: I9a4b22ef08900ef1a4efb52d21e07aefbf1611f1