Update the margin popup adter a LookAndFeel change.
The margin popup creates a JPanel with some controls that is kept
in memory without a parent. The LAF changes will not be propagated
to the popup controls.

This CL fixes the visuals by recreating the margin popup when we
detect a LookAndFeel change since the controls were last created.

Fixes: 131855036
Test: Visual only.
Change-Id: I51a75f6e3fc11e19bcee7656517cf5cea860bac8