[PSD] editable scope bug fix
If we switch viewing a dependency to another dependency of the same
kind (e.g. Library to Library, or Module to Module) we retain the
panel and update the UI elements.  This means that the UI events that
we are using to apply changes to our model do not apply, so we have to
update the configuration explicitly before updating the panel's UI
element.

Since the UI events don't fire, it's not correct to use
getSelectedItem() (because if the user has typed but not performed an
action, the selected item has not been updated).  Since we want the
text anyway, use getEditor().getItem() instead.

Bug: 127954783
Test: manual, existing
Change-Id: I101beda2ee6da612ff5f09999a344a1ad5d64107
(cherry picked from commit 2d943e43b39ffc87ab87e9a0610c154119772dd6)