Don't show perpetual loading screen for unsupported device
In Studio 3.3 we fixed a loading screen bug in 0efafc9, which introduced a bug where
an unsupported device will cause the profiler to show the loading screen forever
when launched from the Profile button (the session dropdown is already disabled and
shows the unsupported reason).

In this CL we detect if the current device has a unsupported message and explicitly
switches to the null stage instead of the loading screen. Also refactored the null
stage to be able to display the unsupported reason if applicable.

Bug: 131926891
Test: StudioProfilerViewTest & StudioProfilersTest
Change-Id: I649a6d1a1d7e8a478282da33cdbe7dce8968ffc1