Add DEPLOYS_TO_LOCAL_DEVICE key for flexibility.
DeviceAndSnapshotComboBoxAction disables itself if the selected run
configuration doesn't deploy to a local device.  This check is done via
matching object types and is not flexible.

This CL adds a key for whether or not a run configuration deploys to
a local device called "android.deploy.DeploysToLocalDevice" that any run
configuration can use to specify if they deploy to a local device.

Bug: 131909473
Test: DeviceAndSnapshotComboBoxActionTest updated.
Change-Id: I6899d8bfd028dadf8546f462462042c84c01ff9f