Handle disposed schemas in NavigationSchema.get
There is a race between a module being disposed and another
thread trying to retrieve the schema for that module.
Instead of throwing if a schema is not available in the
schemas collection, return an empty schema object.
Also add exception handling for registering the schema
as a disposable child of a module, since the module
may have been disposed.

Fixes: 131904224
Test: NavigationSchemaTest.testMissingModule
Change-Id: I255145e79e699f51a56eb6bc2508a42d708c29bc