Make the design area not focusable.
The design area should not be part of the tab order, since <tab>
would change the selection and the properties panel would always
be changing the last selectable component on the design surface.

Also remove the extra border drawn when focus the deisgn area has
focus.

Note: it is still possible to set focus on the sub pane inside the
design area that holds all the components and use <tab> to change
selection. But focus has to be moved into the sub pane with a
mouse click first. And then the focus would stay here (cannot <tab>
out).

Bug: 132077210
Test: N/A
Change-Id: I0ca9165971cfaa7323e54d2bc80b0f31fb6157c0
(cherry picked from commit 441eae7825ef796ab3d936d2da9278893901af2b)