Work around IJ bug where selected setting is no longer valid
There is a bug in IJ that when temporary/shared RunnerAndConfigurationSettings
are cleared and the selected config is part of the removed ones, the selected
config will not get updated and point to an already-removed config. This causes
the device dropdown execution target provider to assert in one of its callees.

This workaround just doublechecks that the selected config is still valid.

Fixes: 131710063
Test: N/A
Change-Id: I3a376e4899a0dd4fdec9f1668af7a315b0e7baff