Check that the selected JDK version matches
... with what the IDE is using.

We are seeing exceptions when sync fails if the selected JDK does not
match to what Android Studio is using.  This is caused because
serialization and de serialization between Gradle and Android Studio.

This CL makes sure that only JDKs with the same version as used by
Studio can be selected from the PSD and from the JDK setup step (which
is currently disabled).

Bug: 124680713
Test: Added tests to confirm only matching Java versions are
considered valid.
Change-Id: I56904a8d9d9998f45a4014b00c8608ed6eb07d37