Extract design tools tracking to the common module
We were tracking design tools actions from the uibuilder module,
probably because some actions being logged are specific to it, for
instance "drop from pallete".

Also, our glue code, i.e. "if thisState then log(thatAction)", was
spread across the code base, which makes it hard to maintain.

To fix this issue, we first extracted the logic of logging studio events
to the common module, in the new CommonUsageTracker interface. The old
NlUsageTracker still handles actions specific to UI builder. This way,
actions being tracked from the common module shouldn't refer to a tracker
located in the uibuilder package.

Another thing done by this CL is to remove the glue code from the
application code to another layer (called *AnalyticsManager), which is
responsible for interfacing the design surface with the corresponding
action tracker, handling the logic of converting the application state
into trackable types.

Bug: N/A
Test: designer and naveditor unit tests

Change-Id: I9c0886fde3a9d9bde997697f73159d93d82f2273