Add flag and entry point for CpuCaptureStage
Creating the CaptureStage allows us to sandbox some of the new
UI components and make changes with some isolation. The flag
is only meant to be for a small transition. This is because
there are large amounts of code that can be removed from the
CpuProfilerStage and CpuProfilerStageView when we make a full
transition.

Test: Added CpuProfilerStage to validate flag, and empty CaptureStage
Bug: 132268755
Change-Id: I5c98648c02209bce7234fe802f22192db059e67b