Show warning for JAVA_HOME only when applicable
We show a warning when users JDK location does not match JAVA_HOME and
suggest to use it (with a quickfix), but if JAVA_HOME is pointing to a
JDK location with an incompatible version, then users will not be able
to have that in the PSD and the quickfix will cause serializaion issues
between Studio and Gradle.

Test: Manually checked that the warning is not shown when using a
different version.
Bug: 124680713

Change-Id: Ibb3b371bd1fb614bd8b44d942e4fb51bcb8e8fbd