[PSD] don't override cursor in dependencies view
The dependencies view has support for hyperlinks, with a particular
cursor when a hyperlink is active.  That's fine, but we have to be
careful not to set a cursor when a hyperlink is not active, otherwise
we will override cursor settings from our parents (e.g. resize cursors
from the parent splitter)

Test: manual
Fixes: 131758184
Change-Id: I4991388e75a82e00220f1301b0b1c2f6c524199e