Fixes issue where multiple tabs are created
When launches have the following pattern: Debug, Run, Run, Run...
Each subsequent Run after the Debug will cause more tabs to be created
under the Run tool window. This is caused by a check to ensure that the
ProcessHandlers' executors match, for IRv1 purposes. This CL removes this
check, since that's not needed any longer.

This fix more or less follows the fix in master.

Fixes: 132444983
Test: verify this is no longer the case
Change-Id: I52bac94c007c831c5806c868f8d8be3ab22d9c01