[resui] Module dropdown: switch to Swing combobox
To enable dynamic resizing of the combobox and search
field, replace the IJ action toolbar for a vanilla swing component.

This caused the dropdown to not update every 500ms like before and thus
was not updated when the selected module was changed.

Test: Manually tested, unit test for this is too complicated
(need to create a second test module)
Bug: None

Change-Id: I2b4b160087273add88a7bace76b3dc50b13a1ccf
(cherry picked from commit aaca1abd38208a12dc7a08cfacfbb23ab8d869f0)