Change WNA resource filename to be unversioned
Change the packaged resource files from 3.4.0.xml, 3.5.0.xml, etc. to
just whats-new-assistant.xml. We only need 1 config bundled with a
certain version of Studio, so it doesn't make sense to have an xml for
3.4.0 in Studio 3.5.0.

Cached xml remains versioned to allow quicker checking, otherwise we
would have to parse the cached file to know whether it's an incorrect
version, which might happen when Studio is updated.

Bug: 132290423
Bug: 130367627
Test: WhatsNewAssistantBundleTest and all existing WNA tests
Change-Id: I4044aca33c648f8a2dff43487eff4c8c6ce9a1bd