Refactor Cpu tracing to follow standard data workflow
Previously cpu traces do not get saved directly into the database
from the data poller. Instead, the profiler model receives the trace details
from the stop-tracing response, manipulates it as needed before calling
a rpc to insert the info into the database. This workflow is different
from most other scenarios, where data flow from daemon -> database -> profilers,
and it will not be compatible with the new pipeline.

By moving the logic to the standard data flow, the cpu profiler now receives
all incoming trace infos (in-progress and completed) and reactively select
and parse a trace whenever it sees a newly completed trace. The trace contents
are made available via the standard byte cache to eliminate the redundant
GetTrace rpc. The new workflow lost the abilities to determine and indicate the
actual capture duration and threads in the CpuTraceInfo. The compromise is
deemed okay as those information are available once a capture is selected and
parsed.

Bug: 119261457
Test: Added and updated

Change-Id: Ifbf6e1f89605f3c98dcc6ef9bfc459aec4a8ca3e