[PSD] don't automatically select resolved dependencies
The Resolved Dependencies panel in the Dependencies PSD view holds a
tree of depenendencies, which was used in Android Studio 3.4 to show
the resolved dependencies corresponding to the currently-selected
Declared Dependency.  This third duty for the tree (in addition to
showing all the resolved dependencies, and allowing them to be
explicitly selected to select the corresponding declared dependency)
is problematic firstly because the resolved dependencies are typically
very far apart from each other (separated by variant and by all the
other dependencies) and secondly because of focus-stealing issues.

Remove the automatic selection of resolved dependencies; start the
resolved dependencies tree initially collapsed (to assist the user in
finding the part of the tree they're interested in); and tidy up
selection handling.

Bug: 127954783 (#13)
Test: manual
Change-Id: I700ba22fd3988a75eea387b73470ced4683dd2e2
(cherry picked from commit 9d168442c26ad28b75a603913c74a1300c9ab2de)