studio analytics: disable when running in IntelliJ
Disable studio analytics when running in Intellij.
We were already not sending metrics when running as a plugin to IntelliJ
but in debug builds we were throwing errors as the analytics code was not
initialized.

Now we explicitly initialize the analytics to a disabled state to allow us
to distinguish between code sending metrics too early vs code sending metrics
when the analytics are disabled.

Fixes: 131213695
Test: manual

Change-Id: Ifa00530578cb8f4ba5e811a10f48d91db54dedd3
(cherry picked from commit f2f82096973e15c635155ca7a2b416736a1806db)