Prints error message if attempting to perform stale swap
When a previously failed swap generates a hyperlink text in the execution
console, this link can be clicked any time afterwards as long as the user
never closes the tool window. After the user changes the state of
Android Studio or device to not match the initial conditions when the
hyperlink was first printed to the console, the hyperlink becomes stale.

This CL forces any apply changes action (the only type that could go stale)
to match the behavior of the button at the moment the user clicks the link.

Fixes: 132359319
Test: Run app, change resource, ACC, kill app, click link, check output
Change-Id: Ic588353f7b07597878c64dd0bbc430ddda1dea1d
(cherry picked from commit 3a49c41b534a350899ed1c7c9d78b0c44ea00062)