Add AndroidModuleSystem#getDirectResourceModuleDependents
For efficient cache invalidation for merged manifests, we'd like to
be able to determine which Android modules transitively depend on a
given Android module for resources. Depending on how the project's build
system models the underlying IntelliJ project, this may result in
processing more dependent modules than is necessary.

This CL resolves the issue by introducing the
getDirectResourceModuleDependents() API to the module system abstracton,
allowing us to define build system-specific strategies for determining
the Android modules that directly depend on a given module for resources.
From this, we can construct build system-independent algorithms for reasoning
over transitive dependents (which I'll add in a follow-up CL).

Gradle and default projects are accurately represented by the IML files,
so the implementations of getDirectResourceModuleDependents() in
GradleModuleSsytem and DefaultModuleSystem just redirect to
ModuleManager#getModuleDependentModules.

Bug: 132629996
Test: Redirects to existing method
Change-Id: I4e86e2d7b3b99d54b492ecc71a0c6eee1ca15902