Introduce MergedManifestRefreshListener
Currently, MergedManifestMananger relies on the assumption that callers
will use the get() API frequently enough to keep the cached merged
manifest snapshot reasonably up to date. This may not be sufficient
if we want to use the merged manifest in light classes, which must
be constructed synchronously and therefore would have to rely on
whatever the MergedManifestManager has cached.

This CL ensures that the cached merged manifest snapshots stay up to
date regardless of caller frequency by introducing
MergedManifestRefreshListener, a project-wide VFS listener that tells
MergedManifestManager to recompute the merged manifest in the background
whenever it detects that a relevant file has changed.

It's worth noting that there could be very many relevant events,
depending on what the user is doing (i.e. if they're directly editing
a manifest file). In such cases, the frequent requests to recompute
will be throttled by the ThrottlingAsyncSupplier backing
MergedManifestManager.

Bug: 132629996
Test: AswbMergedManifestTest
Change-Id: I879e9e958cb88835ddb1b18e65baaa5c0b5b2a42