Fix problem with hit testing for horizontal actions
Currently we add the check for the presence of the
pop icon when the scene component is created. If
an existing action is changed to a pop action, we
would not update the hit provider.

Change to always add the horizontal action hit
provider and check for the pop icon in addHit.

Fixes: 120619257
Test: testHoverGlobalAction
Change-Id: Ib89e4539a77a82837d75fd3d99d3fc199960c958