[PSD] maintain configurations block
In android modules, in some cases (when dependencies have configurations
involving multiple flavour dimensions) we need to explicitly list
configurations used in order for gradle to understand the dependencies
block.

This implementation considers adding a configuration to the
configurations block whenever a dependency configuration is set, and
removes entries from the block only when it is definitely safe to do
so (the configuration properties block is empty, and this is the last
dependency in the module with that configuration).  Configurations are
only considered for addition when there is a chance that this module
will require them (when there is at least one flavorDimension, given
that there are always buildTypes), and are only added if the
configuration contains at least two flavors/buildTypes.

Test: PsAndroidModuleTest.testConfigurationRequiresWorkaround
Test: DependencyManagementTest.testFlavorConfigurationWorkaround
Fixes: 79673293
Change-Id: I84944c4de7051e066503ba9e5e46037e2afda706