Fix "unresolved symbol" with buildSrc module.
Create ContentRootData with BaseGradleProjectResolverExtension
for buildSrc module, which is neccessary to populate buildscript
classpath later.

This CL intentionally keeps the original logic for Android Gradle
projects, for example, JavaModuleModel is still created for buildSrc
module. This was to minimize the possible side effects for 3.5.
For the long term, we should try to remove ModuleModels, use DataNode
and DataService framework to maximize compatibility with Intellij
framework.

Bug: 130138336
Test: included.
Change-Id: Ic084c2121a6431eec58979c8f5b2dca606ed72d6