Show number of resources per category in resource picker
Since the resource picker only expands one category by default is not
very clear that there are more resources in the other categories.

This is specially confusing when adding an ImageView in the layout
editor since the Drawable and Color tabs look exactly the same, it's
hard to realize they each have different resources. This change should
help to clarify that there are more available resources.

Fixes: 123247170
Test: n/a (ui tweak)
Change-Id: I0006e2405a9f80a433b3ab2d83469c4534580a04