Only show debuggers relevant to current project.
Currently the attach debugger dialog shows a list of available debuggers
to attach.  This list of debuggers is populated through various IDE
plugins that may not be compatible with the current project.  E.g.
A debugger that only works for gradle projects should not be an
available option for a non-gradle project.

This CL fixes this by checking AndroidDebugger#supportsProject before
adding the debugger to the list of available debuggers shown to the
user.

Bug: 132949961
Test: Only debuggers compatible to current projects shows up.
Change-Id: I51a7fccc58c831b7bcf472d755fd275144e00eee