Track facets for pending changes in PoliteAndroidVirtualFileListener
This CL updates PoliteAndroidVirtualFileListener so that it determines
and caches the AndroidFacet a VirtualFile belongs to before it is moved
or deleted, and then passes the cached facet to fileChanged() when the
pending move/deletion completes. This simplifies the interface provided by
PoliteAndroidVirtualFileListener so that implementing classes don't have
to consider pending events.

Bug: 134095591
Test: SampleDataResourceRepositoryTest still passes
Change-Id: I8c137fbe7daf82f98e43738d13340df066a72174