Add property delegates to NeleNewPropertyItem
The properties:
 - type
 - componentName
 - libraryName
 - definition (AttributeDefinition)
were not properly delegated to the delegate property for a new
property item. This caused completions and help to be broken for
properties being added in the "Declared Attributes" section of the
properties panel.

Fixes: 133225561
Test: Added checks to existing unit test.
Change-Id: I31e372e954ac8a36af88d6d225056630f1bb0eed
(cherry picked from commit e79c424361c7f0d8c805c7cca5023043b8926d85)