[PSD] more tests for multi-flavor workaround
Test that the configuration entry in the configurations block is
removed when the last dependency with that configuration has its
configuration modified, but not when it is not the last dependency
with that configuration, or when the configuration block is
non-trivial.

(In the process, test also that vanilla configurations don't get added
to the configurations block).

Test: new DependencyManagementTest.testFlavorConfigurationWorkaroundRemoval
Test: (some other tests updated for new psdDependency module)
Bug: 79673293
Change-Id: I3b9e4143b0884e67d757c1ca28eb687991562227