[PSD] detect potentially-colliding buildTypes
Construct the set of all the possible (productFlavor,buildType)
combinations before the new build type is added, and cause a
validation failure if for any productFlavor prefix, the
(productFlavor,newBuildType) is already present in the set.

It might be tempting to try to do something cleverer than this,
perhaps by doing a backwards suffix search along the buildType /
productFlavor dimensions, and terminating (with validation failure) if
we ever have an exact match.  Getting this right is probably too
challenging for the benefit provided, though if anyone complains of
performance problems in adding new build types, this would be the
place to look.

Bug: 113632767
Test: new PsAndroidModuleTest.testValidateBuildTypeNameWithCollisions
Change-Id: I1eb85b87d1df5aa3962468909c0819acce5c22fc