[PSD] detect potentially-colliding product flavors
Product flavors are associated with flavor dimensions, and in general
a given flavor name can be valid or invalid depending on which
dimension it is associated with.  validateProductFlavorName therefore
needs to know which dimension we are attempting to validate a flavor
name for (null for "we don't know", which might happen through some
paths through the UI).

Now that we know which flavor dimension a candidate name is associated
with, we can generate the potential variant name combinations using
specifically that candidate name in the appropriate dimension, and
then perform the same test as for build types to reject product flavor
names which would lead to ambiguous configurations.

Test: PsAndroidModuleTest.testValidateProductFlavorNameWithCollisions
Bug: 113632767
Change-Id: I65f6eb853c03f15ec1715cffec72915f96ced949