Use pipes in ProxySettings exceptions as Gradle
gradle.properties files can have proxy settings and with that they can
include exceptions that are lists of patterns of sites for which the
proxy will not be applied. This list of patterns must be separated by |
but the UI uses a comma separated list.

This CL causes ProxySettings to use | as separator but let the UI as it
is to match other windows provided from Intellij.

Bug: 131991567
Test: Created a couple of tests to check functions that replace
separators
Change-Id: If36ddc5dccf5537810cf41daa0bb6f92bd14dc57