Refactor live allocation tracking workflow in preparation for new pipeline
The BatchAllocationEvents and BatchAllocationContexts samples are now stored and queried
as-is for live-allocation tracking. The profilers model is responsible for unbatching the
samples and look for the allocation events that satisfy the query range. The new workflow
removes some memory optimizaiton where we lazily query for callstack frame details, we can
revisit once the migration is fully complete.

For the legacy pipeline, we need to add buffer to the query range to make sure the last+next
samples are returned that might contain relevant events (This will be solved by the -1/+1
events automatically in the new pipeline).

FakeMemoryService has been updated to return batched data that work with the new workflow.

Bug: 121392346
Test: Existing

Change-Id: I8a86c7167d52cc9c027454f4ad7d3a5febf4f2fd