Fix classloader issues in UI tests on JDK 11
For the reasons described in issue 135142609, the defining ClassLoader
of GuiTestLauncher is no longer an instance of UrlClassLoader in JDK 9+,
so the method previously employed for extracting the classpath for the
client IDE no longer works. A much simpler alternative (using the
contents of the java.class.path system property) appears to work in both
cases.

Also fix finding the java executable for JDK 9+.

Fixes: 135142609
Change-Id: Ie4cc465ccd78f0fb9776bba06bfcb6c14035b946