Teach AndroidModuleSystem about test scopes
This gives us an easy to use API for figuring out which classes and
files are reachable from where, while respecting separate test scopes
when necessary. This logic was previously contained in
TestArtifactResolveScopeProvider which meant it could not be reused for
e.g. finding Room entities visible in a given file.

Bug: None, just a refactoring
Test: existing
Change-Id: Ie2e3c667b592de2d5c4193ef3a894772f5d95794