[PSD] restore selection in All Modules view
When the user performs an action (e.g. by clicking a quickfix link) in
the All Modules dependency view, the entire tree of dependencies is
reset, which clears the existing selection.

It's not completely obvious to me what the Right Thing to do is:
unlike in the Declared Dependencies view, the selection can in
principle include resolved dependencies.  At the moment, because of
the modal nature of the PSD, dependency modifications other than
addition only happen when the declared dependency in question is
selected, but this could change in the future.

Bug: 131766339
Test: manual
Change-Id: Ibb8fd075d722bf1436bbc3765e9c3cea05207be6