[PSD] good names for android dependency versions
What's in a name?  That which we call a library can have arbitrary
characters, but in practice will tend to have lower case letters and
hyphens.  Because we can't yet externalise or round-trip variables
with hyphens in their name, and also to harmonise with the standard
kotlin_version variable, convert hyphens to underscores when proposing
a variable name to the user.

Bug: 135231380
Test: new ExtractVariableWorkerTest.testExtractAndroidModuleDependencyVersion
Change-Id: I15e4d01bac7599436e48ea633074dd335b63ca21