Fix the gradle file merger to recognize an single value assigment with the equal identifier
When a destination gradle file contains a single value assignment with
the equal identifier, it was treated as empty because the
GradleFileSimpleMerger assumed an assignment with the equal identifier
as assignment with a list of values (wrapped with []).

This CL adds a case for an assignment with a single value.

Bug: 135588298
Test: GradleFileSimpleMergerTest
Change-Id: Idbc4006acdb28cae7e93be7c89ee8da698fb692c