Allow loading of native layoutlib behind a flag
This creates a StudioFlag to select whether to use the normal or native
version. Studio needs a restart after changing the value of the flag.
This also modifies the FrameworkResourceRepository so that loads
either the compiled or non-compiled version of the nine-patches
depending on which version of layoutlib is loaded.

Bug: 117921091
Test: Manually checked loading and basic functionality of the native
layoutlib
Change-Id: I6f26fb1fb7c73391b0d702c57a607d985e55372f