Simplify chmod command in TransportFileManager
Previously the TransportFileManager would check for a special
string from the output of chmod to see if there were any errors.
This is error prone, since the string could change, and because
IShellOutputReceiver does not even guarantee that the entire
output of the command will be seen all at once. It seems
easier to just use the exit code of chmod to determine when
to fall back to the old command needed for older devices.

Indeed, the code used to check for the string "Bad mode", and when
using chmod on my device I currently see "bad mode" instead, which
would have been missed by the check.

Bug: n/a
Test: existing
Change-Id: I7bd66181a0b86d2ea446d943f6a48dc4f25517ba