Add safe args gradle plugin to the test dependencies.
Also fixes the bug that the apply directive
(e.g. <apply plugin="androidx.navigation.safeargs.kotlin" />)
appends a new line at the end of the destination build.gradle file.
Which may have a conflict with other plugins such as google services
plugin, which is expected to be at the bottom of the build.gradle file.

Bug: 134632627
Bug: 136634065
Test: TemplateTest, GradleFileMergerTestCase
Change-Id: I8f3e975f98e93db80744d9f7145a911c78478694